A lot of the weapons Reworked + animation. USE THESE IN GAME. NOT THE DRAFTS. More coming Soon
These are all of the reworked weapons that I have made in last weeks. Like I said before. More are coming soon. I will be done with all of the weapons and moving onto to animating a good portion of the enemies by near the end of the week.

Saagar Rai