Updated Douglas and Monsterra
Ok so I have added an in place animation to the monsterra plant. It does look a little basic but seeing how we need a lot of enemies in a very short amount of time, this will have to do. For now. I will try to implement more animations for other enemies. Besides that, I will mostly be done with all of the enemy animations by the end of the week