Evergreen Tree refinement
The reworked pine tree is finally fully completed and I don't have to worry about it anymore, for now. YIPPEE. Besides that, there is now a evergreen tree pallet that we can now use to color evergreen trees. I'll got more to add onto this. If there is anything you need me to do, just message me on discord.